+++ azure/src/main/scala/org/broadinstitute/dsde/workbench/azure/AzureVmServiceInterp.scala
@@ -34,7 +34,9 @@ class AzureVmServiceInterp[F[_]](clientSecretCredential: ClientSecretCredential)
         )
         .map(Option(_))
         .handleErrorWith {
-          case e: ManagementException if e.getValue.getCode().equals("ResourceNotFound") => F.pure(none[VirtualMachine])
+          case e: ManagementException
+              if e.getValue.getCode().equals("ResourceNotFound") | e.getValue.getCode().equals("AuthorizationFailed") =>
+            F.pure(none[VirtualMachine])
           case e => F.raiseError[Option[VirtualMachine]](e)
         }
       res <- tracedLogging(
@@ -59,7 +61,9 @@ class AzureVmServiceInterp[F[_]](clientSecretCredential: ClientSecretCredential)
         )
         .map(Option(_))
         .handleErrorWith {
-          case e: ManagementException if e.getValue.getCode().equals("ResourceNotFound") => F.pure(none[Accepted[Void]])
+          case e: ManagementException
+              if e.getValue.getCode().equals("ResourceNotFound") | e.getValue.getCode().equals("AuthorizationFailed") =>
+            F.pure(none[Accepted[Void]])
           case e => F.raiseError[Option[Accepted[Void]]](e)
         }
       res <- tracedLogging(
@@ -80,7 +84,9 @@ class AzureVmServiceInterp[F[_]](clientSecretCredential: ClientSecretCredential)
       )
       .map(Option(_))
       .handleErrorWith {
-        case e: ManagementException if e.getValue.getCode().equals("ResourceNotFound") => F.pure(none[Mono[Void]])
+        case e: ManagementException
+            if e.getValue.getCode().equals("ResourceNotFound") | e.getValue.getCode().equals("AuthorizationFailed") =>
+          F.pure(none[Mono[Void]])
         case e => F.raiseError[Option[Mono[Void]]](e)
       }
     res <- tracedLogging(
@@ -103,16 +109,18 @@ class AzureVmServiceInterp[F[_]](clientSecretCredential: ClientSecretCredential)
       .delay(
         azureComputeManager
           .virtualMachines()
-          .powerOffAsync(cloudContext.managedResourceGroupName.value, name.value)
+          .deallocateAsync(cloudContext.managedResourceGroupName.value, name.value)
       )
       .map(Option(_))
       .handleErrorWith {
-        case e: ManagementException if e.getValue.getCode().equals("ResourceNotFound") => F.pure(none[Mono[Void]])
+        case e: ManagementException
+            if e.getValue.getCode().equals("ResourceNotFound") | e.getValue.getCode().equals("AuthorizationFailed") =>
+          F.pure(none[Mono[Void]])
         case e => F.raiseError[Option[Mono[Void]]](e)
       }
     res <- tracedLogging(
       fa,
-      s"com.azure.resourcemanager.compute.models.VirtualMachines.powerOffAsync(${cloudContext.managedResourceGroupName.value}, ${name})"
+      s"com.azure.resourcemanager.compute.models.VirtualMachines.deallocateAsync(${cloudContext.managedResourceGroupName.value}, ${name})"
     )
   } yield res
 }