+++ serviceTest/src/test/scala/org/broadinstitute/dsde/workbench/service/RestException.scala
@@ -1,3 +1,10 @@
 package org.broadinstitute.dsde.workbench.service
 
-case class RestException(message: String = null, cause: Throwable = null) extends Exception(message, cause)
+import akka.http.scaladsl.model.StatusCode
+
+case class RestException(message: String = null, cause: Throwable = null, statusCode: StatusCode = null)
+    extends Exception(message, cause)
+
+object RestException {
+  def apply(message: String, statusCode: StatusCode) = new RestException(message, null, statusCode)
+}